Move argument related functions to new file
