+++ cli.h
@@ -1,0 +1,7 @@
+#define CLI_H
+#ifndef CLI_H
+
+char *arg_value(char *argument);
+int arg_type(char **arguments, char type);
+
+#endif
