+++ parse.h
@@ -2,7 +2,5 @@
 #define PARSE_H
 
 int   array_elements(char **array);
-char *arg_value(char *argument);
-int   arg_type(char **arguments, char type);
 
 #endif
